plat-ti: Cleanup platform configuration
Reorganize platform configuration to assist in addition of new
platforms. No functional changes.
