+++ core/arch/arm/plat-ti/main.c
@@ -50,6 +50,7 @@
 static struct gic_data gic_data;
 static struct serial8250_uart_data console_data __early_bss;
 
+register_phys_mem(MEM_AREA_IO_SEC, SECRAM_BASE, SECRAM_SIZE);
 register_phys_mem(MEM_AREA_IO_SEC, GICC_BASE, GICC_SIZE);
 register_phys_mem(MEM_AREA_IO_SEC, GICD_BASE, GICD_SIZE);
 register_phys_mem(MEM_AREA_IO_NSEC, CONSOLE_UART_BASE,
