+++ core/arch/arm/plat-ti/platform_config.h
@@ -28,18 +28,14 @@
 #ifndef PLATFORM_CONFIG_H
 #define PLATFORM_CONFIG_H
 
+#define DRAM0_BASE      0x80000000
+#define DRAM0_SIZE      0x80000000
+
 #if defined(PLATFORM_FLAVOR_dra7xx) || defined(PLATFORM_FLAVOR_am57xx)
 
-#define DRAM0_BASE		0xbdb00000
-#define DRAM0_SIZE		0x02000000
-
-#ifdef CFG_WITH_PAGER
-#error Pager not supported on this platform
-#endif /*CFG_WITH_PAGER*/
-
 /* Location of protected DDR on the DRA7xx platform */
-#define TZDRAM_BASE		0xbdb00000
-#define TZDRAM_SIZE		0x01c00000
+#define TZDRAM_BASE     0xbdb00000
+#define TZDRAM_SIZE     0x01c00000
 
 #define CFG_TEE_CORE_NB_CORE	2
 
@@ -47,20 +43,25 @@
 #define UART2_BASE      0x4806C000
 #define UART3_BASE      0x48020000
 
-/* UART1 */
+#if defined(PLATFORM_FLAVOR_dra7xx)
 #define CONSOLE_UART_BASE       UART1_BASE
+#elif
+#define CONSOLE_UART_BASE       UART3_BASE
+#endif
+
 #define CONSOLE_BAUDRATE        115200
 #define CONSOLE_UART_CLK_IN_HZ	48000000
 
-#define GIC_BASE        0x48210000
+#define SCU_BASE        0x48210000
 #define GICC_OFFSET     0x2000
 #define GICC_SIZE       0x1000
 #define GICD_OFFSET     0x1000
 #define GICD_SIZE       0x1000
-#define GICC_BASE       (GIC_BASE + GICC_OFFSET)
-#define GICD_BASE       (GIC_BASE + GICD_OFFSET)
+#define GICC_BASE       (SCU_BASE + GICC_OFFSET)
+#define GICD_BASE       (SCU_BASE + GICD_OFFSET)
 
 #define SECRAM_BASE     0x40200000
+#define SECRAM_SIZE     0x00100000
 
 /* RNG */
 #define RNG_BASE        0x48090000
@@ -69,41 +70,32 @@
 #error "Unknown platform flavor"
 #endif
 
-#if defined(PLATFORM_FLAVOR_am57xx)
-
-/* UART3 */
-#undef CONSOLE_UART_BASE
-#define CONSOLE_UART_BASE       UART3_BASE
-
-#endif
-
 /* Make stacks aligned to data cache line length */
 #define STACK_ALIGNMENT		64
-
-/* Full GlobalPlatform test suite requires CFG_SHMEM_SIZE to be at least 2MB */
-#define CFG_SHMEM_START		(DRAM0_BASE + TZDRAM_SIZE)
-#define CFG_SHMEM_SIZE		0x400000
-
-#define CFG_TEE_RAM_VA_SIZE	(1024 * 1024)
-
-#ifndef CFG_TEE_LOAD_ADDR
-#define CFG_TEE_LOAD_ADDR	(CFG_TEE_RAM_START + 0x100)
-#endif
 
 /*
  * Assumes that either TZSRAM isn't large enough or TZSRAM doesn't exist,
  * everything is in TZDRAM.
- * +------------------+
- * |        | TEE_RAM |
- * + TZDRAM +---------+
+ * +--------+---------+
+ * |  DRAM  |  SHMEM  |
+ * +--------+---------+
  * |        | TA_RAM  |
+ * | TZDRAM +---------+
+ * |        | TEE_RAM |
  * +--------+---------+
  */
-#define CFG_TEE_RAM_PH_SIZE	CFG_TEE_RAM_VA_SIZE
-#define CFG_TEE_RAM_START	TZDRAM_BASE
+#define CFG_TEE_RAM_VA_SIZE     (1 * 1024 * 1024)
+#define CFG_TEE_RAM_PH_SIZE     CFG_TEE_RAM_VA_SIZE
+#define CFG_TEE_RAM_START       TZDRAM_BASE
+#define CFG_TEE_LOAD_ADDR       (CFG_TEE_RAM_START + 0x100)
+
 #define CFG_TA_RAM_START	ROUNDUP((TZDRAM_BASE + CFG_TEE_RAM_VA_SIZE), \
 					CORE_MMU_DEVICE_SIZE)
 #define CFG_TA_RAM_SIZE		ROUNDDOWN((TZDRAM_SIZE - CFG_TEE_RAM_VA_SIZE), \
 					  CORE_MMU_DEVICE_SIZE)
+
+/* Full GlobalPlatform test suite requires CFG_SHMEM_SIZE to be at least 2MB */
+#define CFG_SHMEM_START         (TZDRAM_BASE + TZDRAM_SIZE)
+#define CFG_SHMEM_SIZE          (4 * 1024 * 1024)
 
 #endif /*PLATFORM_CONFIG_H*/
