+++ scripts/checkpatch_inc.sh
@@ -3,7 +3,7 @@
 CHECKPATCH="${CHECKPATCH:-checkpatch.pl}"
 # checkpatch.pl will ignore the following paths
 CHECKPATCH_IGNORE=$(echo core/lib/lib{fdt,tomcrypt} lib/lib{png,utils,zlib} \
-		core/arch/arm/plat-ti/api_monitor_index.h)
+		core/arch/arm/plat-ti/api_monitor_index_a15.h)
 _CP_EXCL=$(for p in $CHECKPATCH_IGNORE; do echo ":(exclude)$p" ; done)
 
 function _checkpatch() {
